retrieval-agent/requirements.txt
fastapi
uvicorn==0.29.0
faiss-cpu==1.8.0
numpy==1.26.4
sentence-transformers==2.7.0
pydantic==2.7.1